powerpc/64s: move some exception handlers out of line
stable inclusion
from linux-4.19.159
commit b4e80478115bc7fcea627110c133f4afe1a40899
category: bugfix
bugzilla: NA
CVE: CVE-2020-4788

--------------------------------

(backport only)

We're about to grow the exception handlers, which will make a bunch of them
no longer fit within the space available. We move them out of line.

This is a fiddly and error-prone business, so in the interests of reviewability
I haven't merged this in with the addition of the entry flush.

Signed-off-by: Daniel Axtens <dja@axtens.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>
Signed-off-by: Yu Changchun <yuchangchun1@huawei.com>