Document: add guideline to submitting patches to OpenHarmony kernel
ohos inclusion
category: doc
issue: NA
CVE: NA

We need to document our rules of submitting patches to OpenHarmony,
it contains the basic step of submitting patches, and unified patch
format.

This guideline is important for OpenHarmony daily works, so I prefer
to place it into kernel root directory, it would be easy to access and
reference it.

Signed-off-by: Xie XiuQi <xiexiuqi@huawei.com>
Signed-off-by: Hanjun Guo <guohanjun@huawei.com>
[add ohos patch template and tips for developers]
Signed-off-by: Yu Changchun <yuchangchun1@huawei.com>
Signed-off-by: Zachery Wu <wuzuoyu@huawei.com>